rename classes to match java naming convention
