+++ wls-ergebnismeldung-service/src/main/java/de/muenchen/oss/wahllokalsystem/ergebnismeldungservice/service/ergebnismeldung/validation/electionType/BtwValidationImpl.java
@@ -15,13 +15,13 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class EUW_ValidationImpl implements ElectionTypeValidation {
+public class BtwValidationImpl implements ElectionTypeValidation {
 
     private final DefaultElectionTypeValidator validator;
 
     @Override
     public boolean supports(final WahlartModel wahlart) {
-        return WahlartModel.EUW == wahlart;
+        return WahlartModel.BTW == wahlart;
     }
 
     @Override
@@ -35,18 +35,22 @@ public class EUW_ValidationImpl implements ElectionTypeValidation {
     public boolean isValidBwb(final String wahlbezirkID, final String wahlID, final Long waehlerverzeichnisNummer, final MeldungsartModel meldungsart)
             throws WlsException {
         val necessaryStacks = buildNecessaryStack();
-        necessaryStacks.add(Stapelart.EUW_B_LEER);
-
+        necessaryStacks.add(Stapelart.BTW_C_LEER);
         return validator.checkValidation(WahlbezirkArtModel.BWB, wahlbezirkID, wahlID, waehlerverzeichnisNummer, necessaryStacks);
     }
 
     private List<Stapelart> buildNecessaryStack() {
         List<Stapelart> necessaryStacks = new ArrayList<>();
-        necessaryStacks.add(Stapelart.EUW_A);
-        necessaryStacks.add(Stapelart.EUW_B_UNGEKENNZEICHNET);
-        necessaryStacks.add(Stapelart.EUW_C_GUELTIG);
-        necessaryStacks.add(Stapelart.EUW_C_UNGUELTIG);
+        necessaryStacks.add(Stapelart.BTW_A);
+        necessaryStacks.add(Stapelart.BTW_B_I_GUELTIG);
+        necessaryStacks.add(Stapelart.BTW_B_I_UNGUELTIG);
+        necessaryStacks.add(Stapelart.BTW_B_II_GUELTIG);
+        necessaryStacks.add(Stapelart.BTW_B_II_UNGUELTIG);
+        necessaryStacks.add(Stapelart.BTW_C_UNGEKENNZEICHNET);
+        necessaryStacks.add(Stapelart.BTW_D_I_GUELTIG);
+        necessaryStacks.add(Stapelart.BTW_D_II_GUELTIG);
+        necessaryStacks.add(Stapelart.BTW_D_I_UNGUELTIG);
+        necessaryStacks.add(Stapelart.BTW_D_II_UNGUELTIG);
         return necessaryStacks;
     }
-
 }
