+++ wls-ergebnismeldung-service/src/main/java/de/muenchen/oss/wahllokalsystem/ergebnismeldungservice/service/ergebnismeldung/validation/electionType/ObwValidationImpl.java
@@ -15,13 +15,12 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class MBW_ValidationImpl implements ElectionTypeValidation {
+public class ObwValidationImpl implements ElectionTypeValidation {
 
     private final DefaultElectionTypeValidator validator;
 
-    @Override
     public boolean supports(final WahlartModel wahlart) {
-        return WahlartModel.MBW == wahlart;
+        return WahlartModel.OBW == wahlart;
     }
 
     @Override
@@ -35,14 +34,17 @@ public class MBW_ValidationImpl implements ElectionTypeValidation {
     public boolean isValidBwb(final String wahlbezirkID, final String wahlID, final Long waehlerverzeichnisNummer, final MeldungsartModel meldungsart)
             throws WlsException {
         val necessaryStacks = buildNecessaryStack();
+        necessaryStacks.add(Stapelart.OBW_B_LEER);
+
         return validator.checkValidation(WahlbezirkArtModel.BWB, wahlbezirkID, wahlID, waehlerverzeichnisNummer, necessaryStacks);
     }
 
     private List<Stapelart> buildNecessaryStack() {
         List<Stapelart> necessaryStacks = new ArrayList<>();
-        necessaryStacks.add(Stapelart.MBW_A);
-        necessaryStacks.add(Stapelart.MBW_B);
-        necessaryStacks.add(Stapelart.MBW_D_UNGUELTIG);
+        necessaryStacks.add(Stapelart.OBW_A);
+        necessaryStacks.add(Stapelart.OBW_B_UNGEKENNZEICHNET);
+        necessaryStacks.add(Stapelart.OBW_C_GUELTIG);
+        necessaryStacks.add(Stapelart.OBW_C_UNGUELTIG);
         return necessaryStacks;
     }
