+++ wls-ergebnismeldung-service/src/main/java/de/muenchen/oss/wahllokalsystem/ergebnismeldungservice/service/ergebnismeldung/validation/electionType/VeBeValidationImpl.java
@@ -15,12 +15,13 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class OBW_ValidationImpl implements ElectionTypeValidation {
+public class VeBeValidationImpl implements ElectionTypeValidation {
 
     private final DefaultElectionTypeValidator validator;
 
+    @Override
     public boolean supports(final WahlartModel wahlart) {
-        return WahlartModel.OBW == wahlart;
+        return WahlartModel.VE == wahlart || WahlartModel.BEB == wahlart;
     }
 
     @Override
@@ -34,18 +35,17 @@ public class OBW_ValidationImpl implements ElectionTypeValidation {
     public boolean isValidBwb(final String wahlbezirkID, final String wahlID, final Long waehlerverzeichnisNummer, final MeldungsartModel meldungsart)
             throws WlsException {
         val necessaryStacks = buildNecessaryStack();
-        necessaryStacks.add(Stapelart.OBW_B_LEER);
-
         return validator.checkValidation(WahlbezirkArtModel.BWB, wahlbezirkID, wahlID, waehlerverzeichnisNummer, necessaryStacks);
     }
 
     private List<Stapelart> buildNecessaryStack() {
         List<Stapelart> necessaryStacks = new ArrayList<>();
-        necessaryStacks.add(Stapelart.OBW_A);
-        necessaryStacks.add(Stapelart.OBW_B_UNGEKENNZEICHNET);
-        necessaryStacks.add(Stapelart.OBW_C_GUELTIG);
-        necessaryStacks.add(Stapelart.OBW_C_UNGUELTIG);
+        necessaryStacks.add(Stapelart.VE_BE_D_OPTION_1);
+        necessaryStacks.add(Stapelart.VE_BE_D_OPTION_2);
+        necessaryStacks.add(Stapelart.VE_BE_D_UNGUELTIG);
+        necessaryStacks.add(Stapelart.VE_BE_CBA_OPTION_1);
+        necessaryStacks.add(Stapelart.VE_BE_CBA_OPTION_2);
+        necessaryStacks.add(Stapelart.VE_BE_CBA_UNGUELTIG);
         return necessaryStacks;
     }
-
 }
