+++ wls-ergebnismeldung-service/src/test/java/de/muenchen/oss/wahllokalsystem/ergebnismeldungservice/service/ergebnismeldung/validation/electionType/LtwBzwValidationImplTest.java
@@ -29,13 +29,13 @@ import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
-class EUW_ValidationImplTest {
+class LtwBzwValidationImplTest {
 
     @Mock
     DefaultElectionTypeValidator defaultElectionTypeValidator;
 
     @InjectMocks
-    EUW_ValidationImpl unitUnderTest;
+    LtwBzwValidationImpl unitUnderTest;
 
     @Captor
     ArgumentCaptor<List<Stapelart>> captorStapelList;
@@ -44,18 +44,26 @@ class EUW_ValidationImplTest {
     class Supports {
 
         @Test
-        void should_returnTrue_when_wahlartIsEUW() {
-            Assertions.assertThat(unitUnderTest.supports(WahlartModel.EUW)).isTrue();
+        void should_returnTrue_when_wahlartIsLTW() {
+            Assertions.assertThat(unitUnderTest.supports(WahlartModel.LTW)).isTrue();
+        }
+
+        @Test
+        void should_returnTrue_when_wahlartIsBZW() {
+            Assertions.assertThat(unitUnderTest.supports(WahlartModel.BZW)).isTrue();
         }
 
         @ParameterizedTest
-        @MethodSource("argumentsForNonEUWWahlart")
-        void should_returnFalse_when_wahlartIsNotEUW(final ArgumentsAccessor arguments) {
+        @MethodSource("argumentsForNonLTWOrBZWWahlart")
+        void should_returnFalse_when_wahlartIsNotBTW(final ArgumentsAccessor arguments) {
             Assertions.assertThat(unitUnderTest.supports(arguments.get(0, WahlartModel.class))).isFalse();
         }
 
-        public static Stream<Arguments> argumentsForNonEUWWahlart() {
-            return Arrays.stream(WahlartModel.values()).filter(wahlart -> !WahlartModel.EUW.equals(wahlart)).map(Arguments::of);
+        public static Stream<Arguments> argumentsForNonLTWOrBZWWahlart() {
+            return Arrays.stream(WahlartModel.values())
+                    .filter(wahlart -> !WahlartModel.LTW.equals(wahlart))
+                    .filter(wahlart -> !WahlartModel.BZW.equals(wahlart))
+                    .map(Arguments::of);
         }
     }
 
@@ -75,8 +83,10 @@ class EUW_ValidationImplTest {
                     .checkValidation(eq(WahlbezirkArtModel.UWB), eq(wahlbezirkID), eq(wahlID), eq(waehlerverzeichnisNummer), captorStapelList.capture());
 
             val expectedStapel = Arrays.stream(Stapelart.values())
-                    .filter(stapelart -> !Stapelart.EUW_B_LEER.equals(stapelart))
-                    .filter(stapelart -> stapelart.name().startsWith("EUW_"))
+                    .filter(stapelart -> !Stapelart.LTW_BZW_G_KLEIN.equals(stapelart))
+                    .filter(stapelart -> !Stapelart.LTW_BZW_G_GROSS.equals(stapelart))
+                    .filter(stapelart -> !Stapelart.LTW_BZW_G_BEIDE.equals(stapelart))
+                    .filter(stapelart -> stapelart.name().startsWith("LTW_"))
                     .toList().toArray(new Stapelart[0]);
 
             Assertions.assertThat(captorStapelList.getValue()).containsExactlyInAnyOrder(expectedStapel);
@@ -97,6 +107,21 @@ class EUW_ValidationImplTest {
 
             Assertions.assertThat(result).isEqualTo(mockedValidatorResponse);
         }
+
+        @Test
+        void should_notCallDefaultValidatorWithStapelLTW_BZW_DIII_when_meldungsartIsSchnellmeldung() {
+            val wahlbezirkID = "wahlbezirkID";
+            val wahlID = "wahlID";
+            val waehlerverzeichnisNummer = 0L;
+            val meldungsart = MeldungsartModel.V3;
+
+            unitUnderTest.isValidUwb(wahlbezirkID, wahlID, waehlerverzeichnisNummer, meldungsart);
+
+            Mockito.verify(defaultElectionTypeValidator)
+                    .checkValidation(eq(WahlbezirkArtModel.UWB), eq(wahlbezirkID), eq(wahlID), eq(waehlerverzeichnisNummer), captorStapelList.capture());
+
+            Assertions.assertThat(captorStapelList.getValue()).doesNotContain(Stapelart.LTW_BZW_DII);
+        }
     }
 
     @Nested
@@ -115,7 +140,7 @@ class EUW_ValidationImplTest {
                     .checkValidation(eq(WahlbezirkArtModel.BWB), eq(wahlbezirkID), eq(wahlID), eq(waehlerverzeichnisNummer), captorStapelList.capture());
 
             val expectedStapel = Arrays.stream(Stapelart.values())
-                    .filter(stapelart -> stapelart.name().startsWith("EUW_"))
+                    .filter(stapelart -> stapelart.name().startsWith("LTW_"))
                     .toList().toArray(new Stapelart[0]);
 
             Assertions.assertThat(captorStapelList.getValue()).containsExactlyInAnyOrder(expectedStapel);
