+++ wls-ergebnismeldung-service/src/test/java/de/muenchen/oss/wahllokalsystem/ergebnismeldungservice/service/ergebnismeldung/validation/electionType/SrwBawValidationImplTest.java
@@ -29,13 +29,13 @@ import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
-class MBW_ValidationImplTest {
+class SrwBawValidationImplTest {
 
     @Mock
     DefaultElectionTypeValidator defaultElectionTypeValidator;
 
     @InjectMocks
-    MBW_ValidationImpl unitUnderTest;
+    SrwBawValidationImpl unitUnderTest;
 
     @Captor
     ArgumentCaptor<List<Stapelart>> captorStapelList;
@@ -44,18 +44,26 @@ class MBW_ValidationImplTest {
     class Supports {
 
         @Test
-        void should_returnTrue_when_wahlartIsMBW() {
-            Assertions.assertThat(unitUnderTest.supports(WahlartModel.MBW)).isTrue();
+        void should_returnTrue_when_wahlartIsSRW() {
+            Assertions.assertThat(unitUnderTest.supports(WahlartModel.SRW)).isTrue();
+        }
+
+        @Test
+        void should_returnTrue_when_wahlartIsBAW() {
+            Assertions.assertThat(unitUnderTest.supports(WahlartModel.BAW)).isTrue();
         }
 
         @ParameterizedTest
-        @MethodSource("argumentsForNonMBWWahlart")
-        void should_returnFalse_when_wahlartIsNotMBW(final ArgumentsAccessor arguments) {
+        @MethodSource("argumentsForNonSRWOrNonBAWWahlart")
+        void should_returnFalse_when_wahlartIsNotSRWOrBAW(final ArgumentsAccessor arguments) {
             Assertions.assertThat(unitUnderTest.supports(arguments.get(0, WahlartModel.class))).isFalse();
         }
 
-        public static Stream<Arguments> argumentsForNonMBWWahlart() {
-            return Arrays.stream(WahlartModel.values()).filter(wahlart -> !WahlartModel.MBW.equals(wahlart)).map(Arguments::of);
+        public static Stream<Arguments> argumentsForNonSRWOrNonBAWWahlart() {
+            return Arrays.stream(WahlartModel.values())
+                    .filter(wahlart -> !WahlartModel.SRW.equals(wahlart))
+                    .filter(wahlart -> !WahlartModel.BAW.equals(wahlart))
+                    .map(Arguments::of);
         }
     }
 
@@ -75,10 +83,10 @@ class MBW_ValidationImplTest {
                     .checkValidation(eq(WahlbezirkArtModel.UWB), eq(wahlbezirkID), eq(wahlID), eq(waehlerverzeichnisNummer), captorStapelList.capture());
 
             val expectedStapel = Arrays.stream(Stapelart.values())
-                    .filter(stapelart -> !Stapelart.MBW_A_B.equals(stapelart))
-                    .filter(stapelart -> !Stapelart.MBW_D.equals(stapelart))
-                    .filter(stapelart -> !Stapelart.MBW_B_C.equals(stapelart))
-                    .filter(stapelart -> stapelart.name().startsWith("MBW_"))
+                    .filter(stapelart -> !Stapelart.SRW_BAW_A_B.equals(stapelart))
+                    .filter(stapelart -> !Stapelart.SRW_BAW_D.equals(stapelart))
+                    .filter(stapelart -> !Stapelart.SRW_BAW_B_C.equals(stapelart))
+                    .filter(stapelart -> stapelart.name().startsWith("SRW_BAW_"))
                     .toList().toArray(new Stapelart[0]);
 
             Assertions.assertThat(captorStapelList.getValue()).containsExactlyInAnyOrder(expectedStapel);
@@ -117,10 +125,10 @@ class MBW_ValidationImplTest {
                     .checkValidation(eq(WahlbezirkArtModel.BWB), eq(wahlbezirkID), eq(wahlID), eq(waehlerverzeichnisNummer), captorStapelList.capture());
 
             val expectedStapel = Arrays.stream(Stapelart.values())
-                    .filter(stapelart -> !Stapelart.MBW_A_B.equals(stapelart))
-                    .filter(stapelart -> !Stapelart.MBW_D.equals(stapelart))
-                    .filter(stapelart -> !Stapelart.MBW_B_C.equals(stapelart))
-                    .filter(stapelart -> stapelart.name().startsWith("MBW_"))
+                    .filter(stapelart -> !Stapelart.SRW_BAW_A_B.equals(stapelart))
+                    .filter(stapelart -> !Stapelart.SRW_BAW_D.equals(stapelart))
+                    .filter(stapelart -> !Stapelart.SRW_BAW_B_C.equals(stapelart))
+                    .filter(stapelart -> stapelart.name().startsWith("SRW_BAW_"))
                     .toList().toArray(new Stapelart[0]);
 
             Assertions.assertThat(captorStapelList.getValue()).containsExactlyInAnyOrder(expectedStapel);
@@ -142,5 +150,4 @@ class MBW_ValidationImplTest {
             Assertions.assertThat(result).isEqualTo(mockedValidatorResponse);
         }
     }
-
 }
